requirements.txt
streamlit
pillow
torch
torchvision
transformers>=4.28.0
accelerate
safetensors
pandas